+++ aldrin/src/client.rs
@@ -18,11 +18,9 @@ use crate::core::message::{
     DestroyBusListenerResult, DestroyObject, DestroyObjectReply, DestroyObjectResult,
     DestroyService, DestroyServiceReply, DestroyServiceResult, EmitBusEvent, EmitEvent,
     ItemReceived, Message, QueryIntrospection, QueryIntrospectionReply, QueryIntrospectionResult,
-    QueryServiceInfo, QueryServiceInfoReply, QueryServiceInfoResult, QueryServiceVersion,
-    QueryServiceVersionReply, QueryServiceVersionResult, RemoveBusListenerFilter, SendItem,
-    ServiceDestroyed, Shutdown, StartBusListener, StartBusListenerReply, StartBusListenerResult,
-    StopBusListener, StopBusListenerReply, StopBusListenerResult, SubscribeEvent,
-    SubscribeEventReply, Sync, SyncReply, UnsubscribeEvent,
+    RemoveBusListenerFilter, SendItem, ServiceDestroyed, Shutdown, StartBusListener,
+    StartBusListenerReply, StartBusListenerResult, StopBusListener, StopBusListenerReply,
+    StopBusListenerResult, SubscribeEvent, SubscribeEventReply, Sync, SyncReply, UnsubscribeEvent,
 };
 use crate::core::transport::{AsyncTransport, AsyncTransportExt};
 #[cfg(feature = "introspection")]
@@ -30,7 +28,6 @@ use crate::core::TypeId;
 use crate::core::{
     BusListenerCookie, ChannelCookie, ChannelEnd, ChannelEndWithCapacity, Deserialize, ObjectId,
     ProtocolVersion, Serialize, SerializedValue, SerializedValueSlice, ServiceCookie, ServiceId,
-    ServiceInfo,
 };
 use crate::error::{ConnectError, RunError};
 use crate::function_call_map::FunctionCallMap;
@@ -39,8 +36,8 @@ use crate::handle::request::{
     CloseChannelEndRequest, CreateBusListenerRequest, CreateClaimedReceiverRequest,
     CreateClaimedSenderRequest, CreateLifetimeListenerRequest, CreateObjectRequest,
     CreateServiceRequest, DestroyBusListenerRequest, DestroyObjectRequest, DestroyServiceRequest,
-    EmitEventRequest, HandleRequest, QueryServiceInfoRequest, SendItemRequest,
-    StartBusListenerRequest, StopBusListenerRequest, SyncBrokerRequest, SyncClientRequest,
+    EmitEventRequest, HandleRequest, SendItemRequest, StartBusListenerRequest,
+    StopBusListenerRequest, SyncBrokerRequest, SyncClientRequest,
 };
 #[cfg(feature = "introspection")]
 use crate::handle::request::{IntrospectionQueryResult, QueryIntrospectionRequest};
@@ -90,7 +87,6 @@ where
     function_calls: FunctionCallMap,
     services: HashMap<ServiceCookie, mpsc::UnboundedSender<RawCall>>,
     broker_subscriptions: HashMap<ServiceCookie, HashSet<u32>>,
-    query_service_info: SerialMap<oneshot::Sender<Result<ServiceInfo, Error>>>,
     create_channel: SerialMap<CreateChannelData>,
     close_channel_end: SerialMap<CloseChannelEndRequest>,
     claim_channel_end: SerialMap<ClaimChannelEndData>,
@@ -216,7 +212,6 @@ where
             function_calls: FunctionCallMap::new(),
             services: HashMap::new(),
             broker_subscriptions: HashMap::new(),
-            query_service_info: SerialMap::new(),
             create_channel: SerialMap::new(),
             close_channel_end: SerialMap::new(),
             claim_channel_end: SerialMap::new(),
@@ -362,7 +357,7 @@ where
             Message::SubscribeEventReply(msg) => self.msg_subscribe_event_reply(msg),
             Message::UnsubscribeEvent(msg) => self.msg_unsubscribe_event(msg),
             Message::EmitEvent(msg) => self.msg_emit_event(msg),
-            Message::QueryServiceVersionReply(msg) => self.msg_query_service_version_reply(msg)?,
+            Message::QueryServiceVersionReply(_) => todo!(),
             Message::CreateChannelReply(msg) => self.msg_create_channel_reply(msg)?,
             Message::CloseChannelEndReply(msg) => self.msg_close_channel_end_reply(msg)?,
             Message::ChannelEndClosed(msg) => self.msg_channel_end_closed(msg)?,
@@ -383,7 +378,7 @@ where
             Message::AbortFunctionCall(msg) => self.msg_abort_function_call(msg)?,
             Message::QueryIntrospection(msg) => self.msg_query_introspection(msg).await?,
             Message::QueryIntrospectionReply(msg) => self.msg_query_introspection_reply(msg)?,
-            Message::QueryServiceInfoReply(msg) => self.msg_query_service_info_reply(msg)?,
+            Message::QueryServiceInfoReply(_) => todo!(),
 
             Message::Connect(_)
             | Message::ConnectReply(_)
@@ -554,28 +549,6 @@ where
 
     fn msg_emit_event(&self, _msg: EmitEvent) {
         todo!()
-    }
-
-    fn msg_query_service_version_reply(
-        &mut self,
-        msg: QueryServiceVersionReply,
-    ) -> Result<(), RunError<T::Error>> {
-        if self.protocol_version < ProtocolVersion::V1_17 {
-            if let Some(send) = self.query_service_info.remove(msg.serial) {
-                let res = match msg.result {
-                    QueryServiceVersionResult::Ok(version) => Ok(ServiceInfo::new(version)),
-                    QueryServiceVersionResult::InvalidService => Err(Error::InvalidService),
-                };
-
-                let _ = send.send(res);
-                Ok(())
-            } else {
-                Err(RunError::UnexpectedMessageReceived(msg.into()))
-            }
-        } else {
-            // On version >= 1.17, this library never sends QueryServiceVersion.
-            Err(RunError::UnexpectedMessageReceived(msg.into()))
-        }
     }
 
     fn msg_create_channel_reply(
@@ -1024,31 +997,6 @@ where
         Err(RunError::UnexpectedMessageReceived(msg.into()))
     }
 
-    fn msg_query_service_info_reply(
-        &mut self,
-        msg: QueryServiceInfoReply,
-    ) -> Result<(), RunError<T::Error>> {
-        if self.protocol_version >= ProtocolVersion::V1_17 {
-            if let Some(send) = self.query_service_info.remove(msg.serial) {
-                let res = match msg.result {
-                    QueryServiceInfoResult::Ok(value) => {
-                        Ok(value.deserialize().map_err(RunError::Deserialize)?)
-                    }
-
-                    QueryServiceInfoResult::InvalidService => Err(Error::InvalidService),
-                };
-
-                let _ = send.send(res);
-                Ok(())
-            } else {
-                Err(RunError::UnexpectedMessageReceived(msg.into()))
-            }
-        } else {
-            // On version < 1.17, this library never sends QueryServiceInfo.
-            Err(RunError::UnexpectedMessageReceived(msg.into()))
-        }
-    }
-
     async fn handle_request(&mut self, req: HandleRequest) -> Result<(), RunError<T::Error>> {
         match req {
             HandleRequest::HandleCloned => self.req_handle_cloned(),
@@ -1060,7 +1008,6 @@ where
             HandleRequest::CallFunction(req) => self.req_call_function(req).await?,
             HandleRequest::CallFunctionReply(req) => self.req_call_function_reply(req).await?,
             HandleRequest::EmitEvent(req) => self.req_emit_event(req).await?,
-            HandleRequest::QueryServiceInfo(req) => self.req_query_service_info(req).await?,
             HandleRequest::CreateClaimedSender(req) => self.req_create_claimed_sender(req).await?,
             HandleRequest::CreateClaimedReceiver(req) => {
                 self.req_create_claimed_receiver(req).await?
@@ -1230,31 +1177,6 @@ where
             })
             .await
             .map_err(Into::into)
-    }
-
-    async fn req_query_service_info(
-        &mut self,
-        req: QueryServiceInfoRequest,
-    ) -> Result<(), RunError<T::Error>> {
-        let serial = self.query_service_info.insert(req.reply);
-
-        if self.protocol_version >= ProtocolVersion::V1_17 {
-            self.t
-                .send_and_flush(QueryServiceInfo {
-                    serial,
-                    cookie: req.cookie,
-                })
-                .await
-                .map_err(Into::into)
-        } else {
-            self.t
-                .send_and_flush(QueryServiceVersion {
-                    serial,
-                    cookie: req.cookie,
-                })
-                .await
-                .map_err(Into::into)
-        }
     }
 
     async fn req_create_claimed_sender(