+++ broker/src/broker/test.rs
@@ -36,7 +36,7 @@ async fn disconnect_during_function_call() {
 
     // client2 calls a function on client1 and disconnects before client1 replies.
     let mut client2 = broker.add_client().await;
-    let proxy = Proxy::new(client2.clone(), svc.id()).await.unwrap();
+    let proxy = Proxy::new(&client2, svc.id()).await.unwrap();
     let reply = proxy.call(0, &());
     mem::drop(reply);
     client2.join().await;
@@ -83,7 +83,7 @@ async fn drop_conn_before_function_call() {
     mem::drop(conn1_fut);
 
     // Calling a function on conn1 must not deadlock, but be immediately replied to with an error.
-    let proxy = Proxy::new(client2.clone(), svc.id()).await.unwrap();
+    let proxy = Proxy::new(&client2, svc.id()).await.unwrap();
     let res = time::timeout(Duration::from_millis(500), proxy.call(0, &()))
         .await
         .unwrap();